backend/agents/demand.py
"""Demand Forecasting Agent – analyzes sales trends and seasonal patterns."""

from agent_framework.azure import AzureOpenAIResponsesClient
from azure.identity import AzureCliCredential

from backend.config import PROJECT_ENDPOINT, DEPLOYMENT_NAME
from backend.tools.sales_tools import (
    compare_year_over_year,
    get_sales_summary_by_month,
    get_seasonal_forecast,
    get_top_products,
    query_sales_history,
)

DEMAND_INSTRUCTIONS = """You are the Demand Forecasting Agent for a German confectionery manufacturer.

Your expertise:
- Analyzing historical sales data and identifying trends
- Seasonal demand forecasting (Easter and Christmas are the two peak seasons)
- Year-over-year comparison and growth analysis
- Channel performance analysis (retail, wholesale, online)
- Regional demand patterns across Germany (Nord, Süd, West, Ost) and Export

Key context about this business:
- Products include: Schokolade (chocolate), Fruchtgummi (gummy candy), Marzipan, Saisonware (seasonal items like Osterhasen/Easter bunnies, Weihnachtsmänner/Santa figures, Adventskalender), and Geschenkpackungen (gift boxes)
- Easter season: Jan-Apr (production ramps Jan, peaks Mar)
- Christmas season: Sep-Dec (production ramps Sep, peaks Nov-Dec)
- Chocolate sales dip in summer months
- The company has been growing ~6% year-over-year

When answering:
- Use the available tools to query actual sales data — never guess numbers
- Provide specific quantities and revenue figures
- Highlight notable trends and anomalies
- Give actionable recommendations for production planning
- Use German product names naturally (they are the official product names)
- Be concise and data-driven
"""


def create_demand_agent():
    """Create the demand forecasting agent with tools."""
    client = AzureOpenAIResponsesClient(
        project_endpoint=PROJECT_ENDPOINT,
        deployment_name=DEPLOYMENT_NAME,
        credential=AzureCliCredential(),
    )

    return client.as_agent(
        name="DemandForecaster",
        instructions=DEMAND_INSTRUCTIONS,
        tools=[
            query_sales_history,
            get_sales_summary_by_month,
            compare_year_over_year,
            get_seasonal_forecast,
            get_top_products,
        ],
    )
